refactor(cast): inline log helpers and tests implementation
