+++ crates/cast/src/lib.rs
@@ -16,7 +16,6 @@ use alloy_consensus::{
 use alloy_dyn_abi::{DynSolType, DynSolValue, Specifier};
 use alloy_eips::Encodable2718;
 use alloy_ens::NameOrAddress;
-use alloy_json_rpc::RpcError;
 use alloy_network::{AnyNetwork, BlockResponse, Network};
 use alloy_primitives::{
     Address, B256, I256, Keccak256, LogData, Selector, TxHash, U64, U256, hex,
@@ -24,8 +23,7 @@ use alloy_primitives::{
 };
 use alloy_provider::{Provider, network::eip2718::Decodable2718};
 use alloy_rlp::{Decodable, Encodable};
-use alloy_rpc_types::{BlockId, BlockNumberOrTag, Filter, FilterBlockOption, Log};
-use alloy_transport::TransportErrorKind;
+use alloy_rpc_types::BlockId;
 use base::{Base, NumberWithBase};
 use chrono::DateTime;
 use eyre::{Context, ContextCompat, OptionExt, Result};
@@ -39,7 +37,6 @@ use foundry_common::{
 };
 use foundry_config::Chain;
 use foundry_evm::core::bytecode::InstIter;
-use futures::{FutureExt, StreamExt, TryStreamExt, future::Either};
 #[cfg(feature = "optimism")]
 use op_alloy_consensus as _;
 
@@ -47,13 +44,11 @@ use rayon::prelude::*;
 use serde::Serialize;
 use std::{
     fmt::Write,
-    io,
     marker::PhantomData,
     path::PathBuf,
     str::FromStr,
     sync::atomic::{AtomicBool, Ordering},
 };
-use tokio::signal::ctrl_c;
 
 pub use foundry_evm::*;
 
@@ -99,47 +94,6 @@ impl<P: Provider<N> + Clone + Unpin, N: Network> Cast<P, N> {
     pub const fn new(provider: P) -> Self {
         Self { provider, _phantom: PhantomData }
     }
-}
-
-/// Renders a log as an indented list item.
-fn pretty_log(log: &impl UIfmt) -> String {
-    log.pretty()
-        .replacen('\n', "- ", 1) // Remove empty first line
-        .replace('\n', "\n  ") // Indent
-}
-
-impl<P: Provider<AnyNetwork> + Clone + Unpin> Cast<P, AnyNetwork> {}
-
-/// Returns `true` if `err` is a provider range/result-size limit that retrying over a smaller
-/// range can fix. Network, auth, rate-limit, and malformed-response errors return `false`.
-fn is_range_limit_error(err: &RpcError<TransportErrorKind>) -> bool {
-    // Only HTTP 413 (payload too large) is fixable by a smaller range; other transport errors
-    // (network, auth 401/403, rate-limit 429) are not.
-    if let RpcError::Transport(kind) = err {
-        return kind.as_http_error().is_some_and(|http| http.status == 413);
-    }
-
-    // Range/result-size limits are reported as JSON-RPC server error responses; every other
-    // variant falls through to `false`.
-    let RpcError::ErrorResp(payload) = err else { return false };
-    let message = payload.message.to_ascii_lowercase();
-
-    // Phrases providers use for range/result-size limits, kept specific so rate-limit/quota
-    // wording (e.g. "no more than 10 requests per second") doesn't match.
-    const RANGE_LIMIT_HINTS: &[&str] = &[
-        "block range",
-        "blocks range",
-        "range is too",
-        "range too",
-        "returned more than",
-        "response size",
-        "result set",
-        "too many results",
-        "too many blocks",
-        "maximum block range",
-        "max block range",
-    ];
-    RANGE_LIMIT_HINTS.iter().any(|hint| message.contains(hint))
 }
 
 impl<P: Provider<N>, N: Network> Cast<P, N>
@@ -1723,135 +1677,6 @@ fn explorer_client(
     }
 
     builder.build().map_err(Into::into)
-}
-
-/// Tests for the `eth_getLogs` chunking/bisection helpers, kept in a separate module so they can
-/// use the provider-based [`Cast`] (the `tests` module aliases `Cast` to `SimpleCast`).
-#[cfg(test)]
-mod logs_bisecting {
-    use super::Cast;
-    use alloy_json_rpc::{RequestPacket, ResponsePacket, SerializedRequest};
-    use alloy_network::AnyNetwork;
-    use alloy_provider::ProviderBuilder;
-    use alloy_rpc_client::RpcClient;
-    use alloy_rpc_types::{Filter, Log};
-    use alloy_transport::{
-        TransportError, TransportFut,
-        mock::{Asserter, MockTransport},
-    };
-    use std::{
-        sync::{Arc, Mutex},
-        task::{Context, Poll},
-    };
-    use tower::Service;
-
-    fn log_at(block: u64) -> Log {
-        Log { block_number: Some(block), ..Default::default() }
-    }
-
-    /// Mock transport that records the `eth_getLogs` `[fromBlock, toBlock]` ranges it is asked for
-    /// while delegating the actual responses to a FIFO [`Asserter`].
-    #[derive(Clone)]
-    struct RecordingTransport {
-        inner: MockTransport,
-        ranges: Arc<Mutex<Vec<(String, String)>>>,
-    }
-
-    impl RecordingTransport {
-        fn new(asserter: Asserter) -> Self {
-            Self { inner: MockTransport::new(asserter), ranges: Arc::new(Mutex::new(Vec::new())) }
-        }
-
-        fn record(&self, req: &SerializedRequest) {
-            if req.method() != "eth_getLogs" {
-                return;
-            }
-            let Some(params) = req.params() else { return };
-            let Ok(value) = serde_json::from_str::<serde_json::Value>(params.get()) else { return };
-            let Some(filter) = value.get(0) else { return };
-            let field =
-                |name| filter.get(name).and_then(|v| v.as_str()).unwrap_or_default().to_string();
-            self.ranges.lock().unwrap().push((field("fromBlock"), field("toBlock")));
-        }
-    }
-
-    impl Service<RequestPacket> for RecordingTransport {
-        type Response = ResponsePacket;
-        type Error = TransportError;
-        type Future = TransportFut<'static>;
-
-        fn poll_ready(&mut self, cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-            self.inner.poll_ready(cx)
-        }
-
-        fn call(&mut self, req: RequestPacket) -> Self::Future {
-            match &req {
-                RequestPacket::Single(req) => self.record(req),
-                RequestPacket::Batch(reqs) => reqs.iter().for_each(|req| self.record(req)),
-            }
-            self.inner.call(req)
-        }
-    }
-
-    // A range-limit failure splits depth-first into [0,1]/[2,3] and aggregates in range order.
-    #[tokio::test]
-    async fn bisects_failed_range_and_aggregates_in_order() {
-        let asserter = Asserter::new();
-        asserter.push_failure_msg("query returned more than 10000 results");
-        asserter.push_success(&vec![log_at(0)]);
-        asserter.push_success(&vec![log_at(2)]);
-
-        let transport = RecordingTransport::new(asserter);
-        let ranges = transport.ranges.clone();
-        let provider = ProviderBuilder::<_, _, AnyNetwork>::default()
-            .connect_client(RpcClient::new(transport, true));
-
-        let logs =
-            crate::cmd::logs::get_logs_bisecting(&provider, &Filter::new(), 0, 3).await.unwrap();
-        let blocks: Vec<_> = logs.iter().map(|l| l.block_number).collect();
-        assert_eq!(blocks, vec![Some(0), Some(2)]);
-
-        // The original range fails, then bisection requests exactly the two halves in order.
-        let ranges = ranges.lock().unwrap();
-        assert_eq!(
-            *ranges,
-            vec![
-                ("0x0".to_string(), "0x3".to_string()),
-                ("0x0".to_string(), "0x1".to_string()),
-                ("0x2".to_string(), "0x3".to_string()),
-            ]
-        );
-    }
-
-    // A single-block failure can't be split, so the error is surfaced.
-    #[tokio::test]
-    async fn surfaces_single_block_failure() {
-        let asserter = Asserter::new();
-        asserter.push_failure_msg("query returned more than 10000 results");
-
-        let provider =
-            ProviderBuilder::<_, _, AnyNetwork>::default().connect_mocked_client(asserter);
-
-        let err = crate::cmd::logs::get_logs_bisecting(&provider, &Filter::new(), 5, 5)
-            .await
-            .unwrap_err();
-        assert!(err.to_string().contains("more than 10000 results"), "got: {err}");
-    }
-
-    // A non-range error fails after one request instead of bisecting.
-    #[tokio::test]
-    async fn does_not_bisect_non_range_errors() {
-        let asserter = Asserter::new();
-        asserter.push_failure_msg("unauthorized: invalid api key");
-
-        let provider =
-            ProviderBuilder::<_, _, AnyNetwork>::default().connect_mocked_client(asserter);
-
-        let err = crate::cmd::logs::get_logs_bisecting(&provider, &Filter::new(), 0, 3)
-            .await
-            .unwrap_err();
-        assert!(err.to_string().contains("unauthorized"), "got: {err}");
-    }
 }
 
 #[cfg(test)]
